там некторые проблемы были , убрал

+++ src/main/java/andrey/dev/model/Account/AccountService.java
@@ -7,6 +7,7 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 
 import java.util.HashMap;
+import java.util.NoSuchElementException;
 
 @Service
 public class AccountService {
@@ -23,23 +24,29 @@ public class AccountService {
         this.userService = userService;
     }
 
-    public Account createAccount(int userId ) {
-        try{
+    public Account createAccount(int userId) {
+        try {
             userService.searchUserByUserId(userId);
             Account account = new Account(concurrentAccountId, userId, startMoneyAmount);
-            mapAccount.put(concurrentAccountId++,account);
+            mapAccount.put(concurrentAccountId++, account);
             userService.searchUserByUserId(userId).addToAccountList(account);
             return account;
-        } catch (IllegalArgumentException illegalArgumentException){
-            throw new  IllegalArgumentException() ;
+        } catch (IllegalArgumentException illegalArgumentException) {
+            throw new IllegalArgumentException();
         }
     }
 
     public void depositOnAccount(int accountId, double count) {
+        if (!mapAccount.containsKey(accountId)) {
+            throw new NoSuchElementException();
+        }
         mapAccount.get(accountId).addMoneyToMoneyAmount(count);
     }
 
     public void withdrawFromAccount(int accountId, double count) {
+        if (mapAccount.isEmpty() || !mapAccount.containsKey(accountId)) {
+            throw new NoSuchElementException();
+        }
         if (count >= mapAccount.get(accountId).getMoneyAmount()) {
             mapAccount.get(accountId).DeleteMoneyFromMoneyAmount(count);
         } else {
@@ -48,6 +55,9 @@ public class AccountService {
     }
 
     public void transferMoneyFromAccountToAccount(int sourceId, int recipientID, double count) {
+        if (mapAccount.size() == 0 || !mapAccount.containsKey(sourceId) || !mapAccount.containsKey(recipientID)) {
+            throw new NoSuchElementException();
+        }
         if (count >= mapAccount.get(sourceId).getMoneyAmount()) {
             withdrawFromAccount(sourceId, count);
             depositOnAccount(recipientID, (count * (1 - commissionForTransfer)));
@@ -57,11 +67,14 @@ public class AccountService {
     }
 
     public void closeAccount(int accountId, int userId) {
-        try{
+        if (!mapAccount.containsKey(accountId) || userService.searchUserByUserId(userId).getAccountList().size() == 1) {
+            throw new NoSuchElementException();
+        }
+        try {
             userService.searchUserByUserId(userId).removeFromAccountantList(accountId);
             mapAccount.remove(accountId);
-        } catch (IllegalArgumentException illegalArgumentException){
-            throw  new IllegalArgumentException();
+        } catch (IllegalArgumentException illegalArgumentException) {
+            throw new IllegalArgumentException();
         }
     }
